update odps deploy scripts (#69)
* update odps deploy scripts

* reset version update

* add support for incompatible restore for partitioned variables

* update code style

+++ pai_jobs/deploy_ext.sh
@@ -1,13 +1,27 @@
-curr_path=`readlink -f $0`
+greadlink=`which greadlink 2>/dev/null`
+if [ -n "$greadlink" ]  && [ -e "$greadlink" ]
+then
+  curr_path=`greadlink -f $0`
+else
+  curr_path=`readlink -f $0`
+  if [ -z "$curr_path" ]
+  then
+    echo "for mac: brew install coreutils"
+    exit 1
+  fi
+fi
 curr_dir=`dirname $curr_path`
 root_dir=`dirname $curr_dir`
 
 VERSION=""
 ODPSCMD=odpscmd
-resource_only=0
+# 0: deploy resources and xflow
+# 1: deploy resources only
+# 2: generate resources and xflow only, not deploy
+mode=0
 odps_config=""
 
-while getopts 'V:C:Oc:' OPT; do
+while getopts 'V:C:OGc:' OPT; do
     case $OPT in
         V)
             VERSION="$OPTARG";;
@@ -16,10 +30,13 @@ while getopts 'V:C:Oc:' OPT; do
         c)
             odps_config="$OPTARG";;
         O)
-            resource_only=1;;
+            mode=1;;
+        G)
+            mode=2;;
         ?)
             echo "Usage: `basename $0` -V VERSION [-C odpscmd_path] [-c odps_config_path] [-O]"
             echo " -O: only update easy_rec resource file"
+            echo " -G: generate resource file and xflow, but not deploy"
             echo " -c: odps_config file path"
             echo " -C: odpscmd file path, default to: odpscmd, so in default odpscmd must be in PATH"
             echo " -V: algorithm version, chars must be in [0-9A-Za-z_-]"
@@ -34,21 +51,29 @@ then
 fi
 
 ODPSCMD=`which $ODPSCMD`
-if [ $? -ne 0 ]
+if [ $? -ne 0 ] && [ $mode -ne 2 ]
 then
    echo "$ODPSCMD is not in PATH"
    exit 1
 fi
 
-if [ ! -e "$odps_config" ]
+if [ ! -e "$odps_config" ] && [ $mode -ne 2 ]
 then
-  echo "$odps_config does not exist"
+  if [ -z "$odps_config" ]
+  then
+      echo "odps_config is not set"
+  else
+      echo "odps_config[$odps_config] does not exist"
+  fi
   exit 1
 fi
-odps_config=`readlink -f $odps_config`
+if [ -e "$odps_config" ]
+then
+  odps_config=`readlink -f $odps_config`
+fi
 
 cd $root_dir
-bash scripts/gen_proto.sh
+sh scripts/gen_proto.sh
 if [ $? -ne 0 ]
 then
   echo "generate proto file failed"
@@ -58,39 +83,56 @@ fi
 cd $curr_dir
 
 RES_PATH=easy_rec_ext_${VERSION}_res.tar.gz
-ln -s $root_dir/easy_rec ./
+
+if [ ! -e easy_rec ]
+then
+  ln -s $root_dir/easy_rec ./
+fi
 cp easy_rec/__init__.py easy_rec/__init__.py.bak
 sed -i -e "s/\[VERSION\]/$VERSION/g" easy_rec/__init__.py
 find -L easy_rec -name "*.pyc" | xargs rm -rf
-ln -s ../requirements.txt ./
-tar -cvzhf $RES_PATH easy_rec run.py requirements.txt
+tar -cvzhf $RES_PATH easy_rec run.py
 mv easy_rec/__init__.py.bak easy_rec/__init__.py
-${ODPSCMD} --config=$odps_config -e "add file $RES_PATH -f;"
-if [ $? -ne 0 ]
+
+# 2 means generate only
+if [ $mode -ne 2 ]
 then
-  echo "add $RES_PATH failed"
-  exit 1
+  ${ODPSCMD} --config=$odps_config -e "add file $RES_PATH -f;"
+  if [ $? -ne 0 ]
+  then
+    echo "add $RES_PATH failed"
+    exit 1
+  fi
 fi
-if [ $resource_only -gt 0 ]
+
+# deploy resource only
+if [ $mode -eq 1 ]
 then
   echo "add $RES_PATH succeed, version=${VERSION}"
   echo "[WARNING] will not update xflow"
   echo "   your must specify -Dversion=${VERSION} when run pai -name easy_rec_ext"
   exit 0
 fi
-#rm -rf $RES_PATH
 
 cd easy_rec_flow_ex
 sed -i -e "s/parameter name=\"version\" use=\"optional\" default=\"[0-9A-Za-z_-]\+\"/parameter name=\"version\" use=\"optional\" default=\"$VERSION\"/g" easy_rec_ext.xml
 tar -cvzf easy_rec_flow_ex.tar.gz easy_rec_ext.lua  easy_rec_ext.xml
-cd ../xflow-deploy
-package=../easy_rec_flow_ex/easy_rec_flow_ex.tar.gz
-python xflow_deploy.py conf=${odps_config} package=$package
-if [ $? -ne 0 ]
+
+if [ $mode -ne 2 ]
 then
-   echo "deploy $package failed"
-   exit 1
+  cd ../xflow-deploy
+  package=../easy_rec_flow_ex/easy_rec_flow_ex.tar.gz
+  python xflow_deploy.py conf=${odps_config} package=$package
+  if [ $? -ne 0 ]
+  then
+     echo "deploy $package failed"
+     exit 1
+  else
+     echo "deploy $package succeed"
+  fi
+  rm -rf ../easy_rec_flow_ex/easy_rec_flow_ex ../easy_rec_flow_ex/easy_rec_flow_ex.tar.gz
 else
-   echo "deploy $package succeed"
+  cd $curr_dir
+  echo "Generated RESOURCES: $RES_PATH"
+  echo "Generated XFLOW: easy_rec_flow_ex/easy_rec_flow_ex.tar.gz"
 fi
-rm -rf ../easy_rec_flow_ex/easy_rec_flow_ex ../easy_rec_flow_ex/easy_rec_flow_ex.tar.gz
